Add random color and darkening functionality
As requested in the extra credit portion of this project,
I've made it so each time your mouse passes over a grid square
it will produces a random color. Also, the number of passes over
each grid square is tracked and subtracted for the lightness of
the color which will turn the square black at 10 interactions.